+++ packages/components/slider/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @nextui-org/slider
 
+## 2.2.17
+
+### Patch Changes
+
+- Updated dependencies [[`4c01d1824`](https://github.com/nextui-org/nextui/commit/4c01d1824d4dde22d89232968a3a4c48fe04678f)]:
+  - @nextui-org/tooltip@2.0.41
+
+## 2.2.16
+
+### Patch Changes
+
+- Updated dependencies [[`f36df4362`](https://github.com/nextui-org/nextui/commit/f36df4362f572e8e233d4357f43600265cd5b8d5)]:
+  - @nextui-org/shared-utils@2.0.8
+  - @nextui-org/tooltip@2.0.40
+  - @nextui-org/react-utils@2.0.17
+
 ## 2.2.15
 
 ### Patch Changes